+++ src/test/java/learn/agile/demos/pages/DemoPage.java
@@ -43,23 +43,23 @@ public class DemoPage {
 	}
 
 	/**
-	 * @return results
+	 * @return divide results
 	 * @throws InterruptedException
 	 */
-	public int getCalculatorResults() throws InterruptedException {
+	public float getCalculatorResults() throws InterruptedException {
 		WebElement resultEl = (new WebDriverWait(driver, 300))
 				.until(ExpectedConditions.presenceOfElementLocated(By.id("calculator-result")));
 		Thread.sleep(100); // timing - for some reason on some OS's webdriver returns too quickly;
-		return Integer.parseInt(resultEl.getText());
+		return Float.parseFloat(resultEl.getText());
 	}
 
 	/**
 	 * @param x
 	 * @param y
 	 */
-	public void divideNumbers(final int x, final int y) {
-		driver.findElement(By.id("first-number")).sendKeys(Integer.toString(x));
-		driver.findElement(By.id("second-number")).sendKeys(Integer.toString(y));
+	public void divideNumbers(final float x, final float y) {
+		driver.findElement(By.id("first-number")).sendKeys(Float.toString(x));
+		driver.findElement(By.id("second-number")).sendKeys(Float.toString(y));
 		driver.findElement(By.id("divider-button")).click();
 
 	}
